refactor: move usecases/infras under server/infras
Infra clients (Kubernetes, local Hermeum config) are shared across
usecases and routers, so they belong at the server level rather than
nested inside usecases.
