+++ apps/dashboard/src/server/infras/kubernetes/types/hermes-agent.ts
